subhandler class moved to a new file and included algorithm functions
